+++ packages/hamburger-react/src/Hamburger.test.tsx
@@ -1,9 +1,7 @@
 import React from "react";
-import { cleanup, render } from "@testing-library/react";
+import { render, screen, fireEvent } from "@testing-library/react";
 import { Hamburger } from ".";
 import { axe } from "jest-axe";
-
-afterEach(cleanup);
 
 describe("Hamburger", () => {
     it("should render to document", () => {
@@ -11,33 +9,33 @@ describe("Hamburger", () => {
     });
 
     it("should get class jkl-hamburger--is-active on click", () => {
-        const { getByTestId } = render(<Hamburger />);
+        render(<Hamburger />);
 
-        const burger = getByTestId("jkl-hamburger");
-        burger.click();
+        const burger = screen.getByTestId("jkl-hamburger");
+        fireEvent.click(burger);
 
         expect(burger).toHaveClass("jkl-hamburger--is-active");
     });
 
     it("should get class jkl-hamburger--negative when specified", () => {
-        const { getByTestId } = render(<Hamburger negative />);
+        render(<Hamburger negative />);
 
-        const component = getByTestId("jkl-hamburger");
+        const component = screen.getByTestId("jkl-hamburger");
         expect(component).toHaveClass("jkl-hamburger--negative");
     });
 
     it("should have class jkl-hamburger--is-active if initialactive is true", () => {
-        const { getByTestId } = render(<Hamburger initialIsActive />);
+        render(<Hamburger initialIsActive />);
 
-        const component = getByTestId("jkl-hamburger");
+        const component = screen.getByTestId("jkl-hamburger");
         expect(component).toHaveClass("jkl-hamburger--is-active");
     });
 
     it("should render call onClick and set is active", () => {
         const fn = jest.fn();
-        const { getByTestId } = render(<Hamburger onClick={fn} />);
-        const burger = getByTestId("jkl-hamburger");
-        burger.click();
+        render(<Hamburger onClick={fn} />);
+        const burger = screen.getByTestId("jkl-hamburger");
+        fireEvent.click(burger);
 
         expect(fn).toHaveBeenCalledTimes(1);
 
@@ -45,8 +43,8 @@ describe("Hamburger", () => {
     });
 
     it("should render have correct description", () => {
-        const { getByTestId } = render(<Hamburger description="max is better than micky d" />);
-        const burger = getByTestId("jkl-hamburger");
+        render(<Hamburger description="max is better than micky d" />);
+        const burger = screen.getByTestId("jkl-hamburger");
 
         expect(burger).toHaveAttribute("aria-label", "max is better than micky d");
     });