+++ packages/message-box-react/src/MessageBox.test.tsx
@@ -1,21 +1,19 @@
 import React from "react";
-import { cleanup, render } from "@testing-library/react";
+import { render, screen } from "@testing-library/react";
 import { InfoMessage, ErrorMessage, SuccessMessage, WarningMessage } from ".";
 import { axe } from "jest-axe";
-
-afterEach(cleanup);
 
 describe("Message boxes", () => {
     [true, false].map((fullWidth) => {
         [InfoMessage, ErrorMessage, SuccessMessage, WarningMessage].map((E) => {
             it("should render message title and content", () => {
-                const { getByText } = render(
+                render(
                     <E fullWidth={fullWidth} title="test">
                         content
                     </E>,
                 );
-                getByText("content");
-                getByText("test");
+                screen.getByText("content");
+                screen.getByText("test");
             });
         });
     });