+++ portal/gatsby-config.js
@@ -51,5 +51,11 @@ module.exports = {
                 component: require.resolve(`./src/components/Layout/Layout.tsx`),
             },
         },
+        {
+            resolve: `gatsby-plugin-layout`,
+            options: {
+              component: require.resolve(`./src/components/Layout/Layout.tsx`),
+            },
+          },
     ],
 };