+++ packages/accordion-react/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.12](https://github.com/fremtind/jokul/compare/@fremtind/jkl-accordion-react@1.4.7...@fremtind/jkl-accordion-react@1.4.12) (2020-07-02)
+
+**Note:** Version bump only for package @fremtind/jkl-accordion-react
+
+
+
+
+
+## 1.4.11 (2020-07-01)
+
+**Note:** Version bump only for package @fremtind/jkl-accordion-react
+
+
+
+
+
+## 1.4.10 (2020-06-29)
+
+**Note:** Version bump only for package @fremtind/jkl-accordion-react
+
+
+
+
+
 ## 1.4.9 (2020-06-26)
 
 **Note:** Version bump only for package @fremtind/jkl-accordion-react