+++ packages/portal-components/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.3.10 (2020-07-01)
+
+**Note:** Version bump only for package @fremtind/jkl-portal-components
+
+
+
+
+
+## 0.3.9 (2020-06-29)
+
+**Note:** Version bump only for package @fremtind/jkl-portal-components
+
+
+
+
+
 ## 0.3.8 (2020-06-26)