+++ packages/select-react/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.1.0](https://github.com/fremtind/jokul/compare/@fremtind/jkl-select-react@2.0.12...@fremtind/jkl-select-react@2.1.0) (2020-07-03)
+
+
+### Features
+
+* screen reader only label for select + width prop ([#1085](https://github.com/fremtind/jokul/issues/1085)) ([0b87de1](https://github.com/fremtind/jokul/commit/0b87de178b0d685ca09967ce3670f74305a1238a))
+
+
+
+
+
+## 2.0.16 (2020-07-01)
+
+**Note:** Version bump only for package @fremtind/jkl-select-react
+
+
+
+
+
+## 2.0.15 (2020-06-29)
+
+**Note:** Version bump only for package @fremtind/jkl-select-react
+
+
+
+
+
 ## 2.0.14 (2020-06-26)
 
 **Note:** Version bump only for package @fremtind/jkl-select-react