+++ packages/text-input/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.0.13 (2020-07-01)
+
+**Note:** Version bump only for package @fremtind/jkl-text-input
+
+
+
+
+
+## 2.0.12 (2020-06-29)
+
+**Note:** Version bump only for package @fremtind/jkl-text-input
+
+
+
+
+
 ## [2.0.11](https://github.com/fremtind/jokul/compare/@fremtind/jkl-text-input@2.0.10...@fremtind/jkl-text-input@2.0.11) (2020-06-26)
 
 **Note:** Version bump only for package @fremtind/jkl-text-input