+++ packages/toggle-switch-react/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.2.20 (2020-07-01)
+
+**Note:** Version bump only for package @fremtind/jkl-toggle-switch-react
+
+
+
+
+
+## 2.2.19 (2020-06-29)
+
+**Note:** Version bump only for package @fremtind/jkl-toggle-switch-react
+
+
+
+
+
 ## 2.2.18 (2020-06-26)
 
 **Note:** Version bump only for package @fremtind/jkl-toggle-switch-react