+++ portal/CHANGELOG.md
@@ -3,6 +3,71 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.8.0](https://github.com/fremtind/jokul/compare/@fremtind/portal@3.7.1...@fremtind/portal@3.8.0) (2020-07-02)
+
+
+### Features
+
+* **portal:** add animation to portal ([fa38aae](https://github.com/fremtind/jokul/commit/fa38aae91d6a85ec017a80df54a987de893fd3d8))
+
+
+
+
+
+## [3.7.1](https://github.com/fremtind/jokul/compare/@fremtind/portal@3.7.0...@fremtind/portal@3.7.1) (2020-07-02)
+
+**Note:** Version bump only for package @fremtind/portal
+
+
+
+
+
+# [3.7.0](https://github.com/fremtind/jokul/compare/@fremtind/portal@3.6.5...@fremtind/portal@3.7.0) (2020-07-01)
+
+
+### Features
+
+* **portal:** add support for dark mode to code blocks ([68437f0](https://github.com/fremtind/jokul/commit/68437f05cefd2de8f30950154d776eacedf20da7))
+
+
+
+
+
+## [3.6.5](https://github.com/fremtind/jokul/compare/@fremtind/portal@3.6.4...@fremtind/portal@3.6.5) (2020-07-01)
+
+**Note:** Version bump only for package @fremtind/portal
+
+
+
+
+
+## [3.6.4](https://github.com/fremtind/jokul/compare/@fremtind/portal@3.6.3...@fremtind/portal@3.6.4) (2020-06-30)
+
+
+### Bug Fixes
+
+* portal/package.json & portal/.snyk to reduce vulnerabilities ([#1060](https://github.com/fremtind/jokul/issues/1060)) ([5b35981](https://github.com/fremtind/jokul/commit/5b35981577282ac7984f26c4e0f0f7814066e739))
+
+
+
+
+
+## [3.6.3](https://github.com/fremtind/jokul/compare/@fremtind/portal@3.6.2...@fremtind/portal@3.6.3) (2020-06-29)
+
+**Note:** Version bump only for package @fremtind/portal
+
+
+
+
+
+## [3.6.2](https://github.com/fremtind/jokul/compare/@fremtind/portal@3.6.1...@fremtind/portal@3.6.2) (2020-06-29)
+
+**Note:** Version bump only for package @fremtind/portal
+
+
+
+
+
 ## [3.6.1](https://github.com/fremtind/jokul/compare/@fremtind/portal@3.6.0...@fremtind/portal@3.6.1) (2020-06-26)
 
 **Note:** Version bump only for package @fremtind/portal